Upgrade and pin tensorflow
This patch upgrades tensorflow from `>=1.12.1` to `==1.13.1`. The codebase is written assuming tensorflow `1.x.x`, Which is why it's better to pin tensorflow to a known working version.

We also set constraints for opencv-python to ensure we get a wheel package for python 3.5.

Numpy needed to be upgraded because bumping tensorflow required a newer version of numpym from what I can tell this upgrade doesn't cause any issues.

`tensorflow_gpu` is upgraded to match `tensorflow`